Install CAcert in to systems certificates store
CAcert was already available on Debian and Ubuntu releases until 2014 but not
at the time of writing and hasn't been available for RedHat family for a
similar timeframe.

This merges the moving parts from my own cacert role.
https://github.com/goetzk/ansible-cacert

Per the merge request (#13) I've made the following changes:
- cert install default has been changed to no
- cacert.yml reformatted

My test was to run the role against Debian, Ubuntu and Centos systems.