Add support for pre configuring openssl
It can be handy to set all the configuration for openssl once (particularly key
length) rather then passing in options to openssl every time so this will
adjust the values modified as part of key generation - key length and
req_distinguished_name.

Arguably lthese variables duplicate those used for key generations in
tasks/create but they serve a different purpose as they are passed in to a
command which could be run with almost arbitrary parameters.